Fix outstanding problems in claiming and add SOCKS5 support. (#8406)
This commit fixes the known problems in claiming: incorrect reports of success, better treatment of error code and improved visibility of what the script is doing. There has been extensive testing against both environments to check that it works. The socks5 proxy support has been integrated and works for both methods of calling the claiming script.

Co-authored-by: Timotej Šiškovič <timotej@netdata.cloud>