Added support for using `/etc/cron.d` for auto-updates. (#9598)
* Add crontab fragment for netdata updater.

This adds a crontab fragment to be put in `/etc/cron.d` on system swhich
do not support either `/etc/cron.daily` or `/etc/periodic/daily`,
allowing for proper auto-updater support on such systems.

The crontab will run upates at 02:57, which was chosen arbitrarily
because it does not conflict with th erun times for any major
distribution's daily scheduled tasks.

* Utilize the crontab fragment for updates on systems that need it.

This adds logic to utilize the crontab fragment added by the previous
commit on systems which need it for auto-updates to work. The installer
will preferentially use `/etc/cron.daily` or `/etc/periodic/daily`
instead if they exist, so this should result in no changes for most
users.

* Remove systemd scheduler detection.

We're not actually using it right now, and it breaks handling of
auto-updates on systemd systems.