Added improved auto-update support. (#9966)
* Add systemd timer unit to handle auto updates.

This adds a systemd timer unit and associated service for running the
updaterscript. This allows better support for auto-updates on systems
that use systemd, removing the need for cron or special shims to handle
periodic tasks.

This will be used in preference to our existing auto-update support on
systems running systemd unless the user overrides this behavior.

* Add an option to the installer to override auto-updater type detection.

This allows users to specify what auto-updater scheduling mechanism to
use. This wil persist the selection to updates as well.

Supported values are:

* systemd': Uses a systemd timer unit and service to handle automatic
  updates.
* 'interval': Uses a script in /etc/cron.daily or /etc/periodic/daily.
* `crontab`: Uses a crontab file in /etc/cron.d.

* Change priority of systemd support.

This way existing users won't get converted, and by default you'll still
get emails on failures.

* Case-normalize the value passed to --auto-update-type.

* Fix incorrect naming of crontab file.

* Fixed function naming.