Log message when page cache sleeps for more than 1 second. (#11314)
If the page cache manager can not reserve the requested number of pages, it sleeps based on an exponential-backoff value. Normally, this value is a couple milliseconds. However, in rare cases the sleep duration can grow up to several seconds. This patch logs an error message whenever the sleep period is more than 1 second.