Indicate what root privileges are needed for in kickstart.sh. (#14314)
* Indicate what root privileges are needed for in kickstart.sh.

Instead of just blindly prompting the user for root privileges when they
are needed in the kickstart script, print a message about _why_ we need
root privileges. This improves transparency for users about why they are
being prompted.

* Fix shellcheck issues.

* Fix message formatting.

* Fix handling of root privilege messages.

We should be printing one per command, not one per word of the command.