Fix TOCTOU race in daemon status file handling. (#19924)
On most UNIX systems, including Linux, macOS, and FreeBSD, the `chmod()`
function will follow symlinks if called on a symlink. This means that
any code that uses `chmod()` to apply permissions to a known filename
can be exploited to apply those same permissions to an arbitrary file on
the system that the user the code is running as could call `chmod()` on
by simply replacing the known file with a symlink pointing to the target
file after the known file is created but before its permissions are
modified.:

Because we are already opening the file in question in this case, we
should instead use `fchmod()` to modify the permissions via the file
descriptor before closing the file, which is not vulnerable to this
race condition.